Blum Hack V1
The first fully working version of the Blum Hack program! The first fully working version of the Blum Hack! In the blum_automation file the main program logic (automation) takes place, in the text module the output of the “greeting” and the method of clearing the console takes place. Also in the text module cypthon is used to speed up the module and compress the size of the compiled file.